+++ NEWS.md
@@ -1,7 +1,7 @@
 rmarkdown 2.3
 ================================================================================
 
-- Added `renumber_footnotes` (and related `file_scope`) options to output format definition. This enables handling of duplicate numeric footnote identifiers (e.g. across bookdown chapters) via the pandoc `--file-scope` option (#1837).
+- Added `file_scope` option to output format definition. This enables handling of duplicate numeric footnote identifiers (e.g. across bookdown chapters) via the pandoc `--file-scope` option (#1837).
 
 - Added the customizable `lang` atrribute to `ioslides_presentation` output (thanks, @jooyoungseo, #1841).